tests: Separate tests for fields creation
